refactor to vuex
