+++ src/components/main/Footer.vue
@@ -1,13 +1,18 @@
 <template>
-  <div id="Footer">
-    <b-navbar toggleable="lg" type="" variant="dark">
-      <b-navbar-brand style="color:white;">Online to Offline Bazaar</b-navbar-brand>
-    </b-navbar>
-
-    <button v-if="activeUser">lalal</button>
-
+  <div id="footer">
+    <div class="fixed-bottom">
+      <p style="margin:0.25em">Offline to Online Bazaar : Abdi ~ Alex ~ Nani</p>
+    </div> 
   </div>
 </template>
+
+<style>
+#footer div{
+  background-color:gray;
+  color:white;
+}
+</style>
+
 
 <script>   
  export default {
